requirements.txt
# FastAPI Core
fastapi==0.116.0
uvicorn==0.35.0
starlette==0.46.2

# Database & ODM
motor==3.7.1
beanie==1.30.0
pymongo==4.13.2

# Authentication & Security
passlib==1.7.4
python-jose==3.5.0
python-multipart==0.0.20
bcrypt==4.3.0

# Pydantic & Validation
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
email_validator==2.2.0

# Configuration
python-decouple==3.8
python-dotenv==1.1.1

# Date & Time
pytz==2025.2
python-dateutil==2.9.0.post0

# Excel/CSV Processing
pandas==1.5.3
openpyxl==3.1.2

# HTTP Client
httpx==0.28.1
requests==2.32.4

# PDF Processing
pdf2image==1.17.0
pillow==11.3.0

# Computer Vision (OpenCV)
opencv-python==4.8.0.76
numpy==1.24.3

# Fingerprint Device
pyzk==0.9

# Core Dependencies (automatically installed with above)
anyio==4.9.0
click==8.2.1
h11==0.16.0
httpcore==1.0.9
idna==3.10
sniffio==1.3.1
typing_extensions==4.14.1
packaging==25.0
certifi==2025.7.9
charset-normalizer==3.4.2
urllib3==2.5.0
